Оптимизируй это часть 2
