--- src/components/SortingFilter/SortingFilter.constants/SortingFilter.constants.ts
@@ -1,8 +1,0 @@
-import { SortOrder } from '../SortingFilter.typings/SortingFilter.typings';
-
-export const sortFilters: SortOrder[] = [
-  SortOrder.POPULAR,
-  SortOrder.LOW_TO_HIGH,
-  SortOrder.HIGH_TO_LOW,
-  SortOrder.TOP_RATED,
-];
