+++ src/components/favorites-empty/favorites-empty.tsx
@@ -1,0 +1,13 @@
+const FavoritesEmpty = () => (
+  <section className="favorites favorites--empty">
+    <h1 className="visually-hidden">Favorites (empty)</h1>
+    <div className="favorites__status-wrapper">
+      <b className="favorites__status">Nothing yet saved.</b>
+      <p className="favorites__status-description">
+        Save properties to narrow down search or plan your future trips.
+      </p>
+    </div>
+  </section>
+);
+
+export default FavoritesEmpty;
