+++ src/components/sorting-filter/sorting-filter.constants/sorting-filter.constants.ts
@@ -1,0 +1,8 @@
+import { SortOrder } from '@typings/sorting-filter/sorting-filter';
+
+export const sortFilters: SortOrder[] = [
+  SortOrder.POPULAR,
+  SortOrder.LOW_TO_HIGH,
+  SortOrder.HIGH_TO_LOW,
+  SortOrder.TOP_RATED,
+];
