+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { store } from '@store/index';
-import App from '@components/App/App';
+import App from '@components/app/app';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
