+++ src/pages/login-page/login-page.tsx
@@ -1,0 +1,104 @@
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router';
+import Header from '@components/header/header';
+import { APP_ROUTES } from '@services/constants';
+import { login } from '@store/actions';
+import { useAppSelector, useAppDispatch } from '@store/hooks';
+import { validatePassword } from '@utils/user/user';
+
+const LoginPage = (): JSX.Element => {
+  const dispatch = useAppDispatch();
+  const navigate = useNavigate();
+  const authorizationStatus = useAppSelector(
+    (state) => state.userSlice.authorizationStatus
+  );
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [submitDisabled, setSubmitDisabled] = useState(true);
+
+  const handleEmailChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setEmail(event.target.value);
+  };
+
+  const handlePasswordChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setPassword(event.target.value);
+  };
+
+  useEffect(() => {
+    if (validatePassword(password)) {
+      setSubmitDisabled(false);
+    } else {
+      setSubmitDisabled(true);
+    }
+  }, [password]);
+
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    dispatch(login({ email: email, password: password }));
+  };
+
+  useEffect(() => {
+    if (authorizationStatus) {
+      navigate(APP_ROUTES.MAIN);
+    }
+  }, [navigate, authorizationStatus]);
+
+  return (
+    <div className="page page--gray page--login">
+      <Header />
+
+      <main className="page__main page__main--login">
+        <div className="page__login-container container">
+          <section className="login">
+            <h1 className="login__title">Sign in</h1>
+            <form
+              className="login__form form"
+              action="#"
+              method="post"
+              onSubmit={handleSubmit}
+            >
+              <div className="login__input-wrapper form__input-wrapper">
+                <label className="visually-hidden">E-mail</label>
+                <input
+                  className="login__input form__input"
+                  type="email"
+                  name="email"
+                  placeholder="Email"
+                  required
+                  onChange={handleEmailChange}
+                />
+              </div>
+              <div className="login__input-wrapper form__input-wrapper">
+                <label className="visually-hidden">Password</label>
+                <input
+                  className="login__input form__input"
+                  type="password"
+                  name="password"
+                  placeholder="Password"
+                  required
+                  onChange={handlePasswordChange}
+                />
+              </div>
+              <button
+                className="login__submit form__submit button"
+                type="submit"
+                disabled={submitDisabled}
+              >
+                Sign in
+              </button>
+            </form>
+          </section>
+          <section className="locations locations--login locations--current">
+            <div className="locations__item">
+              <a className="locations__item-link" href="#">
+                <span>Amsterdam</span>
+              </a>
+            </div>
+          </section>
+        </div>
+      </main>
+    </div>
+  );
+};
+
+export default LoginPage;
