+++ src/store/userSlice.ts
@@ -1,15 +1,22 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { TUserEntity } from '@typings/User/User';
-import { setAuthorizationStatus, setUserData } from './actions';
+import { TUserEntity } from '@typings/user/user';
+import { TPlaceEntity } from '@typings/place-card/place-card';
+import {
+  setAuthorizationStatus,
+  setUserData,
+  fetchFavoriteOffer,
+} from './actions';
 
 type UserState = {
   authorizationStatus: boolean;
   userData: TUserEntity | null;
+  favoriteOffers: TPlaceEntity[];
 };
 
 const initialState: UserState = {
   authorizationStatus: false,
   userData: null,
+  favoriteOffers: [],
 };
 
 const userSlice = createSlice({
@@ -23,6 +30,9 @@ const userSlice = createSlice({
       })
       .addCase(setUserData, (state, action) => {
         state.userData = action.payload;
+      })
+      .addCase(fetchFavoriteOffer.fulfilled, (state, action) => {
+        state.favoriteOffers = action.payload;
       });
   },
 });
